Add files via upload
This is notebook file of an Exploratory Data Visualization of a dataset containing data points from the job market in the Data industry. The dataset encapsulates a rich, multifaceted view of the data industry's job market from 2020 to 2023, providing a comprehensive snapshot of salary trends, job titles, and employment types, along with geographical and company-related dynamics. Given today’s competitive job market, these insights are a valuable resource for analyzing the evolving landscape of data-related careers, encompassing a wide array of roles from 'Data Analyst' to 'Machine Learning Engineer' and positions from entry-level to executive status.